+++ packages/vsce/__tests__/__e2e__/specs/jvmServer.spec.ts
@@ -11,7 +11,6 @@
 
 import { expect, test } from "@playwright/test";
 import {
-  clickTreeNode,
   constants,
   findAndClickText,
   findAndClickTreeItem,