+++ packages/vsce/__tests__/__mocks__/@zowe/zowe-explorer-api.ts
@@ -9,7 +9,7 @@
  *
  */
 
-import { createProfile, fetchAllProfilesMock, getAllProfilesMock, getJesApiMock, getUssApiMock, loadNamedProfileMock, profile } from "../";
+import { createProfile, fetchAllProfilesMock, getAllProfilesMock, getJesApiMock, getUssApiMock, loadNamedProfileMock } from "../";
 
 const zeApi: any = jest.createMockFromModule("@zowe/zowe-explorer-api");